Unify the configuration of ldap and fix some issues
